requirements.txt
anthropic
Pillow
tqdm